+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- feat(loads): add idempotency_key to load requests
 - feat(loads): support inline csv data in table load requests
 
 ## [0.11.0] - 2026-08-12